realidado o request do cadastro e atualização do usuario&cliente

+++ app/Http/Controllers/Api/UserController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Http\Requests\UpdateUserRequest;
+use App\Http\Requests\CreateUserRequest;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -10,43 +12,31 @@ class UserController extends Controller
 {
     public function index(Request $request)
     {
-
-       return User::all() ;
-
+        return User::all();
     }
 
-    public function create(Request $request)
-    {
-        $user = new User();
-
-        $user = $this->setData($user, $request);
-
-        $user->save();
-        
-        return $user;
+    public function create(CreateUserRequest $request)
+    {        
+        return User::create($request->all());      
     }
 
-    public function update(Request $request)
-    {
-        $user = new User();
+    public function update(UpdateUserRequest $request, User $user)
+    {       
+            $user->update($request->all());
+            return $user;
 
-        $user = $user->find($request->id);
-
-        $user = $this->setData($user, $request);
-
-        $user->save();
-
-        return $user;
+            
     }
-
-    public function delete(Request $request)
+    public function delete(Request $request, $id)
     {
-        $user = new User();
+        $user = User::find($id);
 
-        $user = $user->find($request->id);
-     
-        $user->delete();
-
+        if ($user) {
+            $user->delete();
+            return redirect()->route('users.index')->with('success', 'Usuário excluído com sucesso.');
+        } else {
+            return redirect()->route('users.index')->with('error', 'Usuário não encontrado.');
+        }
     }
 
     protected function setData(User $user, Request $request)
@@ -60,8 +50,6 @@ class UserController extends Controller
         $user->senha = $request->senha;
         $user->perfil = $request->perfil;
 
-        return $user;        
+        return $user;
     }
-
-
 }
